Wire up the ViewAllBooksController to the Dispatcher Servlet
Map the new ViewAllBooksController to the Dispatcher Servlet in order
to make the controller accessable through the web application.

Here we wire the controller to url: /viewAllBooks.do

The wiring is put in a separate xml-file named according to the
follwing syntax:
- <something>-servlet.xml

where /something/ is the name used for the (Dispatcher) Servlet in the
 =web.xml= config file.

The <something>-servlet.xml file is put in the /WEB-INF/ directory to
prevent web application users from reading it through their browser.