Session management: Try 3, Use scope 'request' on controller (best)
This is probably the Best way to handle sessions.

Use /Scope to Request/, i.e. dependency injection.
- Use annotation /@Scope("request")/ on controller class.

NOTE: In this commit I had to change Dispatcher-servlet.xml to get rid
of an exception when adding an entry to the cart.
This is just a configuration problem, NOT a problem with the code
itself. See the xml-file details.