Add conversion error handling to catch and report system crashes
This adds code to properly handle wrong input data that does not match
the expected 'data type'.

Note: Is does not check if values are reasonable, just the
format. I.e. trying to enter text where a number is expected, which
would cause type cast to fail in the controller.