Use UrlBasedViewResolver and ResourceBundleViewResolver combined
The real value of using view resolvers comes into play when we
combine (chain) the resolvers.

We configure both resolvers and add an /order/ property to them.

If the first resolver doesn't match the second is tried. In this case
the /properties/ file is still used but can be kept small, only
configuring those views that do not match a w3-file.